Rename functions, add comments and move to operator.go
